refactor: split calibration.py to calibration module
