+++ docs/package/constants.md
@@ -1,0 +1,1 @@
+::: pychilaslasers.constants
