--- docs/package/utils.md
@@ -1,1 +1,0 @@
-::: pychilaslasers.utils
