+++ src/pychilaslasers/constants.py
@@ -3,6 +3,7 @@
 This module defines baud rates, error codes, and command lists used by the library.
 """
 
+
 class Constants:
     """Constants used throughout the PyChilasLasers library."""
 
@@ -10,7 +11,7 @@ class Constants:
     TLM_INITIAL_BAUDRATE = (
         57600  # Initial baudrate the laser is set to when it is powered on.
     )
-    SUPPORTED_BAUDRATES: tuple[int,...] = (
+    SUPPORTED_BAUDRATES: tuple[int, ...] = (
         9600,
         14400,
         19200,
@@ -24,9 +25,9 @@ class Constants:
     )
 
     # ERROR CODES THAT SHOULD TRIGGER A ERROR DIALOG (errors 14 to 23)
-    CRITICAL_ERRORS: tuple[str, ...] = tuple(["E0" + str(x) for x in range(14, 24)] + [
-        "E0" + str(x) for x in range(30, 51)
-    ])
+    CRITICAL_ERRORS: tuple[str, ...] = tuple(
+        ["E0" + str(x) for x in range(14, 24)] + ["E0" + str(x) for x in range(30, 51)]
+    )
 
     # Commands that can be replaced with a semicolon to speed up communication in
     # firmware
